refact: - create random-content file for generateContent/comment functions -remove film and comment function
