--- src/utils/film.js
@@ -1,8 +1,0 @@
-import { getRandomInteger } from './common';
-
-export const generateContent = (contentArray) => {
-
-  const randomIndex = getRandomInteger(0, contentArray.length - 1);
-
-  return contentArray[randomIndex];
-};
